Restructure: extract Timetable trait into timetable.rs (algorithm dispatch becomes a free fn)
