+++ raptor/src/lib.rs
@@ -150,6 +150,7 @@ mod ids;
 mod journey;
 mod label;
 mod time;
+mod timetable;
 
 use crate::algorithm::boarding::BoardingTree;
 use crate::algorithm::label_bag::LabelBag;
@@ -177,6 +178,7 @@ pub use label::Label;
 pub use time::Duration;
 pub use time::SecondOfDay;
 pub use time::Transfers;
+pub use timetable::Timetable;
 
 #[cfg(test)]
 mod test;
@@ -184,205 +186,6 @@ mod test;
 /// Internal round-counter type, used only for indexing label arrays.
 /// User-facing transfer caps are passed as [`Transfers`].
 pub(crate) type K = usize;
-
-/// Models a route-based transit network for the RAPTOR algorithm.
-///
-/// Implement this trait to describe your transit network's topology and
-/// schedule. The algorithm itself is invoked via the
-/// [`Timetable::query`] builder.
-///
-/// Identifiers are dense `u32` indices ([`StopIdx`], [`RouteIdx`],
-/// [`TripIdx`]). Adapters intern from external IDs (e.g. GTFS string IDs)
-/// at construction time.
-///
-/// # Footpaths
-///
-/// The footpath relation returned by [`get_footpaths_from`] does **not**
-/// need to be transitively closed: if you can walk `A → B` and `B → C`,
-/// the algorithm will chain them within a single round, reaching `C`
-/// from `A` with the combined walk time. Footpath relaxation iterates to
-/// a fixed point per round.
-///
-/// Closure can still be useful as an optimisation — pre-closed graphs
-/// have fewer edges to traverse — but it is not a soundness
-/// prerequisite.
-///
-/// # No overtaking within a route
-///
-/// All trips returned by [`get_earliest_trip`] for a given route must
-/// share a stop sequence and pairwise must not overtake. The algorithm
-/// uses a binary search by departure time at intermediate stops, which
-/// is only sound when the trip ordering is monotone at every stop.
-/// Adapters that ingest data with multiple stop patterns or overtaking
-/// should split such groups into separate routes at construction.
-///
-/// [`get_footpaths_from`]: Timetable::get_footpaths_from
-/// [`get_earliest_trip`]: Timetable::get_earliest_trip
-pub trait Timetable {
-    /// Number of stops in this timetable. Stop indices are in `0..n_stops()`.
-    fn n_stops(&self) -> usize;
-    /// Number of routes (post-pattern-splitting). Route indices are in
-    /// `0..n_routes()`.
-    fn n_routes(&self) -> usize;
-
-    /// Returns each route serving the given stop, paired with the *earliest*
-    /// position of `stop` within that route's sequence.
-    ///
-    /// For loop routes where `stop` appears more than once on a route, only
-    /// the smallest position is reported. Each route appears at most once
-    /// in the returned slice.
-    fn get_routes_serving_stop(&self, stop: StopIdx) -> &[(RouteIdx, u32)];
-
-    /// Returns the route's stop sequence from `pos` onwards, inclusive.
-    ///
-    /// Iterating the returned slice with positional offsets gives the
-    /// algorithm `(pos + offset, stop_at_position)` pairs without ambiguity,
-    /// even when a route revisits stops.
-    ///
-    /// Panics if `pos` is out of range for the route.
-    fn get_stops_after(&self, route: RouteIdx, pos: u32) -> &[StopIdx];
-
-    /// Returns the stop at the given position within a route's sequence.
-    ///
-    /// Panics if `pos` is out of range for the route.
-    fn stop_at(&self, route: RouteIdx, pos: u32) -> StopIdx;
-
-    /// Finds the earliest trip on a route departing at or after `at` from
-    /// the stop at the given position within the route's sequence.
-    ///
-    /// `pos` disambiguates which visit of the stop to consider when the route
-    /// revisits it. Returns `None` if no trip departs at or after `at`.
-    fn get_earliest_trip(&self, route: RouteIdx, at: SecondOfDay, pos: u32) -> Option<TripIdx>;
-
-    /// Returns the arrival time of a trip at the given position within its
-    /// route's sequence.
-    fn get_arrival_time(&self, trip: TripIdx, pos: u32) -> SecondOfDay;
-
-    /// Returns the departure time of a trip at the given position within its
-    /// route's sequence.
-    fn get_departure_time(&self, trip: TripIdx, pos: u32) -> SecondOfDay;
-
-    /// Returns all stops directly reachable from the given stop via
-    /// walking (footpaths).
-    ///
-    /// The relation does not need to be transitively closed — the
-    /// algorithm chains walks within a round. See the trait-level docs.
-    fn get_footpaths_from(&self, stop: StopIdx) -> &[StopIdx];
-
-    /// Returns the walking transfer time between two stops.
-    /// The default implementation returns 1 second.
-    fn get_transfer_time(&self, from: StopIdx, to: StopIdx) -> Duration {
-        let (_, _) = (from, to);
-        Duration(1)
-    }
-
-    /// Reports whether the footpath relation is transitively closed —
-    /// that is, whether `A → C` is already a direct edge whenever
-    /// `A → B` and `B → C` are. The default is `false`.
-    ///
-    /// When `true`, the algorithm uses a single-pass `O(E)` footpath
-    /// relaxation per round instead of the multi-source Dijkstra
-    /// fallback (`O(E log V)`). This is a meaningful speedup on
-    /// dense closed graphs (e.g. publisher-curated `transfers.txt`
-    /// files in Berlin / Paris feeds).
-    ///
-    /// **Soundness**: returning `true` when the relation is *not*
-    /// closed will cause the algorithm to miss journeys whose optimal
-    /// path requires chaining direct walks within a round. Only return
-    /// `true` if you know the relation is closed.
-    fn footpaths_are_transitively_closed(&self) -> bool {
-        false
-    }
-
-    /// Start a typestate-builder query. Returns a [`Query`] in the
-    /// [`NeedsDeparture`] state. Call `.from(...).to(...).max_transfers(...)`
-    /// (any order, all optional with defaults), then either
-    /// `.depart_at(...)` for a single-departure query or
-    /// `.depart_in_window(...)` for a range query, then `.run()`.
-    ///
-    /// ```no_run
-    /// # use raptor::{Timetable, SecondOfDay, Duration, StopIdx};
-    /// # fn ex<T: Timetable>(tt: &T, start: StopIdx, end: StopIdx) {
-    /// let journeys = tt
-    ///     .query()
-    ///     .from(start)
-    ///     .to(end)
-    ///     .max_transfers(10)
-    ///     .depart_at(SecondOfDay::hms(9, 0, 0))
-    ///     .run();
-    /// # }
-    /// ```
-    fn query(&self) -> Query<'_, Self, ArrivalTime, NeedsDeparture>
-    where
-        Self: Sized,
-    {
-        Query {
-            tt: self,
-            origins: Endpoints::new(),
-            targets: Endpoints::new(),
-            max_transfers: Transfers(10),
-            mode: NeedsDeparture,
-            _label: std::marker::PhantomData,
-        }
-    }
-
-    /// Like [`Timetable::query`] but with a custom [`Label`] type for
-    /// multi-criterion routing. `Vec<Journey<L>>` and `Vec<RangeJourney<L>>`
-    /// come back from the corresponding `.run()`, with each entry on the
-    /// returned Pareto front a different trade-off across `L`'s criteria.
-    ///
-    /// You only need this if [`ArrivalTime`] (the default) is the wrong
-    /// shape for your problem — e.g. you want to surface a slower route
-    /// with less walking. The bundled [`labels::ArrivalAndWalk`] does
-    /// exactly that. See the [`Label`] trait for what's involved in
-    /// writing your own.
-    ///
-    /// ```no_run
-    /// # use raptor::{Timetable, SecondOfDay, StopIdx};
-    /// # use raptor::labels::ArrivalAndWalk;
-    /// # fn ex<T: Timetable>(tt: &T, start: StopIdx, end: StopIdx) {
-    /// let pareto_front = tt
-    ///     .query_with_label::<ArrivalAndWalk>()
-    ///     .from(start)
-    ///     .to(end)
-    ///     .max_transfers(10)
-    ///     .depart_at(SecondOfDay::hms(9, 0, 0))
-    ///     .run();
-    /// # }
-    /// ```
-    fn query_with_label<L: Label>(&self) -> Query<'_, Self, L, NeedsDeparture>
-    where
-        Self: Sized,
-    {
-        Query {
-            tt: self,
-            origins: Endpoints::new(),
-            targets: Endpoints::new(),
-            max_transfers: Transfers(10),
-            mode: NeedsDeparture,
-            _label: std::marker::PhantomData,
-        }
-    }
-
-    /// Implementation entry point for [`Query::run`] /
-    /// [`Query::run_with_cache`]. Public-but-hidden so the typestate
-    /// builder can dispatch into the algorithm. Don't call this
-    /// directly — use [`Timetable::query`] instead.
-    #[doc(hidden)]
-    fn raptor_with_cache_and_label<L: Label>(
-        &self,
-        cache: &mut RaptorCache<L>,
-        transfers: usize,
-        depart: SecondOfDay,
-        origins: impl IntoEndpoints,
-        targets: impl IntoEndpoints,
-    ) -> Vec<Journey<L>>
-    where
-        Self: Sized,
-    {
-        run_per_call_query(self, cache, transfers, depart, origins, targets)
-    }
-}
 
 /// One entry in a range-query profile: a departure time paired with
 /// the [`Journey`] it produces. Returned by [`Query::run`] /
@@ -454,12 +257,12 @@ where
     T: Timetable + ?Sized,
     L: Label,
 {
-    tt: &'tt T,
-    origins: Endpoints,
-    targets: Endpoints,
-    max_transfers: Transfers,
-    mode: M,
-    _label: std::marker::PhantomData<L>,
+    pub(crate) tt: &'tt T,
+    pub(crate) origins: Endpoints,
+    pub(crate) targets: Endpoints,
+    pub(crate) max_transfers: Transfers,
+    pub(crate) mode: M,
+    pub(crate) _label: std::marker::PhantomData<L>,
 }
 
 // ----- Stage 1: NeedsDeparture — optional inputs and mode transitions -----
@@ -552,7 +355,8 @@ where
     /// [`RaptorCache::for_timetable`] with the same timetable you call
     /// the query on to avoid this.
     pub fn run_with_cache(self, cache: &mut RaptorCache<L>) -> Vec<Journey<L>> {
-        self.tt.raptor_with_cache_and_label(
+        run_per_call_query(
+            self.tt,
             cache,
             self.max_transfers.0 as usize,
             self.mode.at,
@@ -635,13 +439,8 @@ where
             .par_iter()
             .flat_map_iter(|&depart| {
                 let mut cache = pool.checkout();
-                let journeys = tt.raptor_with_cache_and_label(
-                    &mut *cache,
-                    transfers,
-                    depart,
-                    &origins,
-                    &targets,
-                );
+                let journeys =
+                    run_per_call_query(tt, &mut *cache, transfers, depart, &origins, &targets);
                 journeys
                     .into_iter()
                     .map(move |j| RangeJourney { depart, journey: j })
